update
notes wow this is a good read love it,,,, it is wordy.. You explain alot here.  I want to be able to grab the Sr. (10+ year ) experience guy's mind .  He already knows how futures works... he already has dealt with async vs sync.. hes looking for the next level... Reactive.. lets just get on with that as much as possible ?